[MIN-354] refactor runtime
