+++ runtime/src/rpc_tests.rs
@@ -1,4 +1,458 @@
-use super::*;
+use crate::{
+	AccountId, Balance, Block, Controller, Currencies, EnabledUnderlyingAssetsIds, LiquidationPools, LiquidityPools,
+	MinterestCouncilMembership, MinterestOracle, MinterestProtocol, MntToken, Prices, Rate, Runtime, System,
+	WhitelistCouncilMembership, DOLLARS, PROTOCOL_INTEREST_TRANSFER_THRESHOLD,
+};
+use controller::{ControllerData, PauseKeeper};
+use controller_rpc_runtime_api::{
+	runtime_decl_for_ControllerRuntimeApi::ControllerRuntimeApi, BalanceInfo, HypotheticalLiquidityData, PoolState,
+	UserPoolBalanceData,
+};
+use frame_support::pallet_prelude::{DispatchResultWithPostInfo, PhantomData};
+use frame_support::{
+	assert_noop, assert_ok, error::BadOrigin, pallet_prelude::GenesisBuild, parameter_types, traits::OnFinalize,
+};
+use liquidation_pools::LiquidationPoolData;
+use liquidity_pools::{Pool, PoolUserData};
+use minterest_model::MinterestModelData;
+use minterest_primitives::{CurrencyId, Operation, Price};
+use mnt_token_rpc_runtime_api::runtime_decl_for_MntTokenRuntimeApi::MntTokenRuntimeApi;
+use orml_traits::MultiCurrency;
+use pallet_traits::{PoolsManager, PricesManager};
+use prices_rpc_runtime_api::runtime_decl_for_PricesRuntimeApi::PricesRuntimeApi;
+use risk_manager::RiskManagerData;
+use sp_runtime::{traits::Zero, DispatchResult, FixedPointNumber};
+use test_helper::{BTC, DOT, ETH, KSM, MDOT, MNT};
+
+parameter_types! {
+	pub ALICE: AccountId = AccountId::from([1u8; 32]);
+	pub BOB: AccountId = AccountId::from([2u8; 32]);
+	pub CHARLIE: AccountId = AccountId::from([3u8; 32]);
+	pub ORACLE1: AccountId = AccountId::from([4u8; 32]);
+	pub ORACLE2: AccountId = AccountId::from([5u8; 32]);
+	pub ORACLE3: AccountId = AccountId::from([6u8; 32]);
+
+}
+
+struct ExtBuilder {
+	endowed_accounts: Vec<(AccountId, CurrencyId, Balance)>,
+	pools: Vec<(CurrencyId, Pool)>,
+	pool_user_data: Vec<(CurrencyId, AccountId, PoolUserData)>,
+	minted_pools: Vec<(CurrencyId, Balance)>,
+}
+
+impl Default for ExtBuilder {
+	fn default() -> Self {
+		Self {
+			endowed_accounts: vec![
+				// seed: initial assets. Initial MNT to pay for gas.
+				(ALICE::get(), MNT, 100_000 * DOLLARS),
+				(ALICE::get(), DOT, 100_000 * DOLLARS),
+				(ALICE::get(), ETH, 100_000 * DOLLARS),
+				(ALICE::get(), BTC, 100_000 * DOLLARS),
+				(ALICE::get(), KSM, 100_000 * DOLLARS),
+				(BOB::get(), MNT, 100_000 * DOLLARS),
+				(BOB::get(), DOT, 100_000 * DOLLARS),
+				(BOB::get(), ETH, 100_000 * DOLLARS),
+				(BOB::get(), BTC, 100_000 * DOLLARS),
+				(BOB::get(), KSM, 100_000 * DOLLARS),
+				(CHARLIE::get(), MNT, 100_000 * DOLLARS),
+				(CHARLIE::get(), DOT, 100_000 * DOLLARS),
+				(CHARLIE::get(), ETH, 100_000 * DOLLARS),
+				(CHARLIE::get(), BTC, 100_000 * DOLLARS),
+				(CHARLIE::get(), KSM, 100_000 * DOLLARS),
+			],
+			pools: vec![],
+			pool_user_data: vec![],
+			minted_pools: vec![
+				(KSM, 2 * DOLLARS),
+				(DOT, 2 * DOLLARS),
+				(ETH, 2 * DOLLARS),
+				(BTC, 2 * DOLLARS),
+			],
+		}
+	}
+}
+
+impl ExtBuilder {
+	pub fn enable_minting_for_all_pools(mut self, speed: Balance) -> Self {
+		self.minted_pools = vec![(KSM, speed), (DOT, speed), (ETH, speed), (BTC, speed)];
+		self
+	}
+
+	pub fn mnt_enabled_pools(mut self, pools: Vec<(CurrencyId, Balance)>) -> Self {
+		self.minted_pools = pools;
+		self
+	}
+
+	pub fn pool_initial(mut self, pool_id: CurrencyId) -> Self {
+		self.pools.push((
+			pool_id,
+			Pool {
+				total_borrowed: Balance::zero(),
+				borrow_index: Rate::one(),
+				total_protocol_interest: Balance::zero(),
+			},
+		));
+		self
+	}
+
+	pub fn mnt_account_balance(mut self, balance: Balance) -> Self {
+		self.endowed_accounts.push((MntToken::get_account_id(), MNT, balance));
+		self
+	}
+
+	pub fn build(self) -> sp_io::TestExternalities {
+		let mut t = frame_system::GenesisConfig::default()
+			.build_storage::<Runtime>()
+			.unwrap();
+
+		pallet_balances::GenesisConfig::<Runtime> {
+			balances: self
+				.endowed_accounts
+				.clone()
+				.into_iter()
+				.filter(|(_, currency_id, _)| *currency_id == MNT)
+				.map(|(account_id, _, initial_balance)| (account_id, initial_balance))
+				.collect::<Vec<_>>(),
+		}
+		.assimilate_storage(&mut t)
+		.unwrap();
+
+		orml_tokens::GenesisConfig::<Runtime> {
+			endowed_accounts: self
+				.endowed_accounts
+				.into_iter()
+				.filter(|(_, currency_id, _)| *currency_id != MNT)
+				.collect::<Vec<_>>(),
+		}
+		.assimilate_storage(&mut t)
+		.unwrap();
+
+		liquidity_pools::GenesisConfig::<Runtime> {
+			pools: self.pools,
+			pool_user_data: self.pool_user_data,
+		}
+		.assimilate_storage(&mut t)
+		.unwrap();
+
+		controller::GenesisConfig::<Runtime> {
+			controller_dates: vec![
+				(
+					DOT,
+					ControllerData {
+						// Set the timestamp to one, so that the accrue_interest_rate() does not work.
+						last_interest_accrued_block: 1,
+						protocol_interest_factor: Rate::saturating_from_rational(1, 10), // 10%
+						max_borrow_rate: Rate::saturating_from_rational(5, 1000),        // 0.5%
+						collateral_factor: Rate::saturating_from_rational(9, 10),        // 90%
+						borrow_cap: None,
+						protocol_interest_threshold: PROTOCOL_INTEREST_TRANSFER_THRESHOLD,
+					},
+				),
+				(
+					ETH,
+					ControllerData {
+						// Set the timestamp to one, so that the accrue_interest_rate() does not work.
+						last_interest_accrued_block: 1,
+						protocol_interest_factor: Rate::saturating_from_rational(1, 10), // 10%
+						max_borrow_rate: Rate::saturating_from_rational(5, 1000),        // 0.5%
+						collateral_factor: Rate::saturating_from_rational(9, 10),        // 90%
+						borrow_cap: None,
+						protocol_interest_threshold: PROTOCOL_INTEREST_TRANSFER_THRESHOLD,
+					},
+				),
+				(
+					BTC,
+					ControllerData {
+						// Set the timestamp to one, so that the accrue_interest_rate() does not work.
+						last_interest_accrued_block: 1,
+						protocol_interest_factor: Rate::saturating_from_rational(1, 10), // 10%
+						max_borrow_rate: Rate::saturating_from_rational(5, 1000),        // 0.5%
+						collateral_factor: Rate::saturating_from_rational(9, 10),        // 90%
+						borrow_cap: None,
+						protocol_interest_threshold: PROTOCOL_INTEREST_TRANSFER_THRESHOLD,
+					},
+				),
+				(
+					KSM,
+					ControllerData {
+						// Set the timestamp to one, so that the accrue_interest_rate() does not work.
+						last_interest_accrued_block: 1,
+						protocol_interest_factor: Rate::saturating_from_rational(1, 10), // 10%
+						max_borrow_rate: Rate::saturating_from_rational(5, 1000),        // 0.5%
+						collateral_factor: Rate::saturating_from_rational(9, 10),        // 90%
+						borrow_cap: None,
+						protocol_interest_threshold: PROTOCOL_INTEREST_TRANSFER_THRESHOLD,
+					},
+				),
+			],
+			pause_keepers: vec![
+				(DOT, PauseKeeper::all_unpaused()),
+				(ETH, PauseKeeper::all_unpaused()),
+				(BTC, PauseKeeper::all_unpaused()),
+				(KSM, PauseKeeper::all_unpaused()),
+			],
+			whitelist_mode: false,
+		}
+		.assimilate_storage(&mut t)
+		.unwrap();
+
+		minterest_model::GenesisConfig {
+			minterest_model_params: vec![
+				(
+					DOT,
+					MinterestModelData {
+						kink: Rate::saturating_from_rational(8, 10),
+						base_rate_per_block: Rate::zero(),
+						multiplier_per_block: Rate::saturating_from_rational(9, 1_000_000_000), // 0.047304 PerYear
+						jump_multiplier_per_block: Rate::saturating_from_rational(207, 1_000_000_000), // 1.09 PerYear
+					},
+				),
+				(
+					ETH,
+					MinterestModelData {
+						kink: Rate::saturating_from_rational(8, 10),
+						base_rate_per_block: Rate::zero(),
+						multiplier_per_block: Rate::saturating_from_rational(9, 1_000_000_000), // 0.047304 PerYear
+						jump_multiplier_per_block: Rate::saturating_from_rational(207, 1_000_000_000), // 1.09 PerYear
+					},
+				),
+				(
+					BTC,
+					MinterestModelData {
+						kink: Rate::saturating_from_rational(8, 10),
+						base_rate_per_block: Rate::zero(),
+						multiplier_per_block: Rate::saturating_from_rational(9, 1_000_000_000), // 0.047304 PerYear
+						jump_multiplier_per_block: Rate::saturating_from_rational(207, 1_000_000_000), // 1.09 PerYear
+					},
+				),
+			],
+		}
+		.assimilate_storage::<Runtime>(&mut t)
+		.unwrap();
+
+		risk_manager::GenesisConfig {
+			risk_manager_params: vec![
+				(
+					DOT,
+					RiskManagerData {
+						max_attempts: 3,
+						min_partial_liquidation_sum: 100_000 * DOLLARS,
+						threshold: Rate::saturating_from_rational(103, 100),
+						liquidation_fee: Rate::saturating_from_rational(105, 100),
+					},
+				),
+				(
+					ETH,
+					RiskManagerData {
+						max_attempts: 3,
+						min_partial_liquidation_sum: 100_000 * DOLLARS,
+						threshold: Rate::saturating_from_rational(103, 100),
+						liquidation_fee: Rate::saturating_from_rational(105, 100),
+					},
+				),
+				(
+					BTC,
+					RiskManagerData {
+						max_attempts: 3,
+						min_partial_liquidation_sum: 100_000 * DOLLARS,
+						threshold: Rate::saturating_from_rational(103, 100),
+						liquidation_fee: Rate::saturating_from_rational(105, 100),
+					},
+				),
+			],
+		}
+		.assimilate_storage::<Runtime>(&mut t)
+		.unwrap();
+
+		pallet_membership::GenesisConfig::<Runtime, pallet_membership::Instance3> {
+			members: vec![ORACLE1::get().clone(), ORACLE2::get().clone(), ORACLE3::get().clone()],
+			phantom: Default::default(),
+		}
+		.assimilate_storage(&mut t)
+		.unwrap();
+
+		liquidation_pools::GenesisConfig::<Runtime> {
+			phantom: PhantomData,
+			liquidation_pools: vec![
+				(
+					DOT,
+					LiquidationPoolData {
+						deviation_threshold: Rate::saturating_from_rational(1, 10),
+						balance_ratio: Rate::saturating_from_rational(2, 10),
+						max_ideal_balance: None,
+					},
+				),
+				(
+					ETH,
+					LiquidationPoolData {
+						deviation_threshold: Rate::saturating_from_rational(1, 10),
+						balance_ratio: Rate::saturating_from_rational(2, 10),
+						max_ideal_balance: None,
+					},
+				),
+				(
+					BTC,
+					LiquidationPoolData {
+						deviation_threshold: Rate::saturating_from_rational(1, 10),
+						balance_ratio: Rate::saturating_from_rational(2, 10),
+						max_ideal_balance: None,
+					},
+				),
+				(
+					KSM,
+					LiquidationPoolData {
+						deviation_threshold: Rate::saturating_from_rational(1, 10),
+						balance_ratio: Rate::saturating_from_rational(2, 10),
+						max_ideal_balance: None,
+					},
+				),
+			],
+		}
+		.assimilate_storage(&mut t)
+		.unwrap();
+
+		module_prices::GenesisConfig::<Runtime> {
+			locked_price: vec![(MNT, Rate::saturating_from_integer(4))],
+			_phantom: PhantomData,
+		}
+		.assimilate_storage(&mut t)
+		.unwrap();
+
+		mnt_token::GenesisConfig::<Runtime> {
+			mnt_claim_threshold: 0, // disable by default
+			minted_pools: self.minted_pools,
+			_phantom: std::marker::PhantomData,
+		}
+		.assimilate_storage(&mut t)
+		.unwrap();
+
+		let mut ext: sp_io::TestExternalities = t.into();
+		ext.execute_with(|| System::set_block_number(1));
+		ext
+	}
+}
+
+fn pool_balance(pool_id: CurrencyId) -> Balance {
+	Currencies::free_balance(pool_id, &LiquidityPools::pools_account_id())
+}
+fn get_protocol_total_value_rpc() -> Option<BalanceInfo> {
+	<Runtime as ControllerRuntimeApi<Block, AccountId>>::get_protocol_total_value()
+}
+
+fn liquidity_pool_state_rpc(currency_id: CurrencyId) -> Option<PoolState> {
+	<Runtime as ControllerRuntimeApi<Block, AccountId>>::liquidity_pool_state(currency_id)
+}
+
+fn get_utilization_rate_rpc(pool_id: CurrencyId) -> Option<Rate> {
+	<Runtime as ControllerRuntimeApi<Block, AccountId>>::get_utilization_rate(pool_id)
+}
+
+fn get_total_supply_and_borrowed_usd_balance_rpc(account_id: AccountId) -> Option<UserPoolBalanceData> {
+	<Runtime as ControllerRuntimeApi<Block, AccountId>>::get_total_supply_and_borrowed_usd_balance(account_id)
+}
+
+fn get_hypothetical_account_liquidity_rpc(account_id: AccountId) -> Option<HypotheticalLiquidityData> {
+	<Runtime as ControllerRuntimeApi<Block, AccountId>>::get_hypothetical_account_liquidity(account_id)
+}
+
+fn is_admin_rpc(caller: AccountId) -> Option<bool> {
+	<Runtime as ControllerRuntimeApi<Block, AccountId>>::is_admin(caller)
+}
+
+fn get_user_total_collateral_rpc(account_id: AccountId) -> Balance {
+	<Runtime as ControllerRuntimeApi<Block, AccountId>>::get_user_total_collateral(account_id)
+		.unwrap()
+		.amount
+}
+
+fn get_user_borrow_per_asset_rpc(account_id: AccountId, underlying_asset_id: CurrencyId) -> Option<BalanceInfo> {
+	<Runtime as ControllerRuntimeApi<Block, AccountId>>::get_user_borrow_per_asset(account_id, underlying_asset_id)
+}
+
+fn get_user_underlying_balance_per_asset_rpc(account_id: AccountId, pool_id: CurrencyId) -> Option<BalanceInfo> {
+	<Runtime as ControllerRuntimeApi<Block, AccountId>>::get_user_underlying_balance_per_asset(account_id, pool_id)
+}
+
+fn get_unclaimed_mnt_balance_rpc(account_id: AccountId) -> Balance {
+	<Runtime as MntTokenRuntimeApi<Block, AccountId>>::get_unclaimed_mnt_balance(account_id)
+		.unwrap()
+		.amount
+}
+
+fn pool_exists_rpc(underlying_asset_id: CurrencyId) -> bool {
+	<Runtime as ControllerRuntimeApi<Block, AccountId>>::pool_exists(underlying_asset_id)
+}
+
+fn dollars(amount: u128) -> u128 {
+	amount.saturating_mul(Price::accuracy())
+}
+
+fn alice() -> <Runtime as frame_system::Config>::Origin {
+	<Runtime as frame_system::Config>::Origin::signed((ALICE::get()).clone())
+}
+
+fn bob() -> <Runtime as frame_system::Config>::Origin {
+	<Runtime as frame_system::Config>::Origin::signed((BOB::get()).clone())
+}
+
+fn charlie() -> <Runtime as frame_system::Config>::Origin {
+	<Runtime as frame_system::Config>::Origin::signed((CHARLIE::get()).clone())
+}
+
+fn origin_of(account_id: AccountId) -> <Runtime as frame_system::Config>::Origin {
+	<Runtime as frame_system::Config>::Origin::signed(account_id)
+}
+
+fn origin_root() -> <Runtime as frame_system::Config>::Origin {
+	<Runtime as frame_system::Config>::Origin::root()
+}
+
+fn set_oracle_price_for_all_pools(price: u128) -> DispatchResult {
+	let prices: Vec<(CurrencyId, Price)> = EnabledUnderlyingAssetsIds::get()
+		.into_iter()
+		.map(|pool_id| (pool_id, Price::saturating_from_integer(price)))
+		.collect();
+	MinterestOracle::on_finalize(System::block_number());
+	assert_ok!(MinterestOracle::feed_values(origin_of(ORACLE1::get().clone()), prices));
+	Ok(())
+}
+
+fn set_oracle_prices(prices: Vec<(CurrencyId, Price)>) -> DispatchResult {
+	MinterestOracle::on_finalize(System::block_number());
+	assert_ok!(MinterestOracle::feed_values(origin_of(ORACLE1::get().clone()), prices));
+	Ok(())
+}
+
+fn get_all_locked_prices() -> Vec<(CurrencyId, Option<Price>)> {
+	<Runtime as PricesRuntimeApi<Block>>::get_all_locked_prices()
+}
+
+fn get_all_freshest_prices() -> Vec<(CurrencyId, Option<Price>)> {
+	<Runtime as PricesRuntimeApi<Block>>::get_all_freshest_prices()
+}
+
+fn lock_price(currency_id: CurrencyId) -> DispatchResultWithPostInfo {
+	Prices::lock_price(origin_root(), currency_id)
+}
+
+fn unlock_price(currency_id: CurrencyId) -> DispatchResultWithPostInfo {
+	Prices::unlock_price(origin_root(), currency_id)
+}
+
+fn get_mnt_borrow_and_supply_rates(pool_id: CurrencyId) -> (Rate, Rate) {
+	<Runtime as MntTokenRuntimeApi<Block, AccountId>>::get_mnt_borrow_and_supply_rates(pool_id).unwrap()
+}
+
+fn run_to_block(n: u32) {
+	while System::block_number() < n {
+		MinterestProtocol::on_finalize(System::block_number());
+		MinterestOracle::on_finalize(System::block_number());
+		System::set_block_number(System::block_number() + 1);
+	}
+}
+
 #[test]
 fn demo_scenario_n2_without_interest_using_rpc_should_work() {
 	ExtBuilder::default()
@@ -1053,4 +1507,119 @@ fn pool_exists_should_work() {
 		assert_eq!(pool_exists_rpc(DOT), true);
 		assert_eq!(pool_exists_rpc(ETH), false);
 	});
+}
+
+#[test]
+fn whitelist_mode_should_work() {
+	ExtBuilder::default().pool_initial(DOT).build().execute_with(|| {
+		// Set price = 2.00 USD for all pools.
+		assert_ok!(set_oracle_price_for_all_pools(2));
+		System::set_block_number(1);
+		assert_ok!(MinterestProtocol::deposit_underlying(bob(), DOT, dollars(10_000)));
+		System::set_block_number(2);
+
+		assert_ok!(Controller::switch_whitelist_mode(
+			<Runtime as frame_system::Config>::Origin::root()
+		));
+		System::set_block_number(3);
+
+		// In whitelist mode only members of 'WhitelistCouncil' can work with protocols.
+		assert_noop!(
+			MinterestProtocol::deposit_underlying(bob(), DOT, dollars(5_000)),
+			BadOrigin
+		);
+		System::set_block_number(4);
+
+		assert_ok!(WhitelistCouncilMembership::add_member(
+			<Runtime as frame_system::Config>::Origin::root(),
+			BOB::get()
+		));
+		System::set_block_number(5);
+
+		assert_ok!(MinterestProtocol::deposit_underlying(bob(), DOT, dollars(10_000)));
+	})
+}
+
+//------------ Protocol interest transfer tests ----------------------
+
+// Protocol interest should be transferred to liquidation pool after block is finalized
+#[test]
+fn protocol_interest_transfer_should_work() {
+	ExtBuilder::default()
+		.pool_initial(DOT)
+		.pool_initial(ETH)
+		.build()
+		.execute_with(|| {
+			// Set price = 2.00 USD for all pools.
+			assert_ok!(set_oracle_price_for_all_pools(2));
+
+			// Set interest factor equal 0.75.
+			assert_ok!(Controller::set_protocol_interest_factor(
+				origin_root(),
+				DOT,
+				Rate::saturating_from_rational(3, 4)
+			));
+
+			assert_ok!(MinterestProtocol::deposit_underlying(alice(), DOT, dollars(100_000)));
+			assert_ok!(MinterestProtocol::deposit_underlying(alice(), ETH, dollars(100_000)));
+
+			System::set_block_number(10);
+
+			assert_ok!(MinterestProtocol::deposit_underlying(bob(), DOT, dollars(50_000)));
+			assert_ok!(MinterestProtocol::deposit_underlying(bob(), ETH, dollars(70_000)));
+			assert_ok!(MinterestProtocol::enable_is_collateral(bob(), DOT));
+			assert_ok!(MinterestProtocol::enable_is_collateral(bob(), ETH));
+			// exchange_rate = (150 - 0 + 0) / 150 = 1
+			assert_eq!(
+				liquidity_pool_state_rpc(DOT),
+				Some(PoolState {
+					exchange_rate: Rate::one(),
+					borrow_rate: Rate::zero(),
+					supply_rate: Rate::zero()
+				})
+			);
+
+			System::set_block_number(20);
+
+			assert_ok!(MinterestProtocol::borrow(bob(), DOT, dollars(100_000)));
+			assert_eq!(LiquidityPools::pools(DOT).total_protocol_interest, Balance::zero());
+
+			System::set_block_number(1000);
+			assert_ok!(MinterestProtocol::repay(bob(), DOT, dollars(10_000)));
+			assert_eq!(pool_balance(DOT), dollars(60_000));
+			MinterestProtocol::on_finalize(1000);
+			// Not reached threshold, pool balances should stay the same
+			assert_eq!(
+				LiquidityPools::pools(DOT).total_protocol_interest,
+				441_000_000_000_000_000u128
+			);
+
+			System::set_block_number(10000000);
+
+			assert_ok!(MinterestProtocol::repay(bob(), DOT, dollars(20_000)));
+			assert_eq!(pool_balance(DOT), dollars(80_000));
+
+			let total_protocol_interest: Balance = 3_645_120_550_951_706_945_733;
+			assert_eq!(
+				LiquidityPools::pools(DOT).total_protocol_interest,
+				total_protocol_interest
+			);
+
+			let liquidity_pool_dot_balance = LiquidityPools::get_pool_available_liquidity(DOT);
+			let liquidation_pool_dot_balance = LiquidationPools::get_pool_available_liquidity(DOT);
+
+			// Threshold is reached. Transfer total_protocol_interest to liquidation pool
+			MinterestProtocol::on_finalize(10000000);
+
+			let transferred_to_liquidation_pool = total_protocol_interest;
+			assert_eq!(LiquidityPools::pools(DOT).total_protocol_interest, Balance::zero());
+			assert_eq!(
+				LiquidityPools::get_pool_available_liquidity(DOT),
+				liquidity_pool_dot_balance - transferred_to_liquidation_pool
+			);
+			assert_eq!(
+				LiquidationPools::get_pool_available_liquidity(DOT),
+				liquidation_pool_dot_balance + transferred_to_liquidation_pool
+			);
+		});
 }
